added apollo-logger plugin
